Fix: 좌석 입력방식 변경을 반영시키도록 기존 날짜별 좌석 조회 테스트 코드 수정

+++ src/test/kotlin/io/ticketaka/api/concert/application/ConcertSeatServiceTest.kt
@@ -70,7 +70,7 @@ class ConcertSeatServiceTest {
         val result = concertSeatService.getSeats(date)
 
         // then
-        assertEquals(listOf(seatNumber.toInt()), result)
+        assertEquals(listOf(seatNumber), result)
     }
 
     @Test
